Make sure the Manage Users permission is checked in set_own_login_name.
This is only when trying to change the login name of someone else.
The pas.updateLoginName method is protected by the Manage Users permission,
but that is of course not checked in this trusted python code,
so we need to do that ourselves, even though it does not match the name of the function.